Populate resources directory with all type definitions from sync
Successfully ran npm run sync with network access to fetch all type definitions:
- 29 Kotlin type definitions cached in resources/kotlin/
- 23 Java type definitions cached in resources/java/
- Kotlin documentation cached in resources/kotlin-doc.html
- 2 nested types (Map.Entry, MutableMap.MutableEntry) failed due to URL structure

Note: Parsing has some issues to fix - types detected as 'interface' instead of 'class' in some cases, and formatting issues with line breaks. Will address in follow-up commit.

+++ resources/kotlin/kotlin_Short.kt
@@ -1,4 +1,4 @@
 package kotlin
 
-interface Int {
+interface Short {
 }
